requirements.txt
setuptools
wheel
selene
selenium
pytest
pytest-listener
pytest-html
testcontainers
webdriver_manager
allure-pytest
test_recorder
docker-compose
requests
PyHamcrest